Guardado de Imagen, e integración con Cliente

+++ src/main/java/com/talentpool/monolito/repository/IImagenRepository.java
@@ -3,5 +3,5 @@ package com.talentpool.monolito.repository;
 import com.talentpool.monolito.model.Imagen;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface IImagenRepository extends MongoRepository<Imagen, Integer> {
+public interface IImagenRepository extends MongoRepository<Imagen, Long> {
 }
